+++ packages/api/src/trueppm_api/apps/projects/serializers.py
@@ -79,6 +79,7 @@ class TaskSerializer(serializers.ModelSerializer[Task]):
             "duration",
             "percent_complete",
             "notes",
+            "planned_start",
             "early_start",
             "early_finish",
             "late_start",