+++ packages/api/src/trueppm_api/apps/sync/serializers.py
@@ -53,6 +53,7 @@ class SyncTaskSerializer(serializers.ModelSerializer[Task]):
             "duration",
             "percent_complete",
             "notes",
+            "planned_start",
             "early_start",
             "early_finish",
             "late_start",